Update package name and move SourceInfoMacro
