+++ core/src/main/scala/chisel3/Num.scala
@@ -4,7 +4,7 @@ package chisel3
 
 import chisel3.internal.ChiselException
 import chisel3.internal.firrtl.{BinaryPoint, KnownBinaryPoint}
-import chisel3.internal.sourceinfomacros.SourceInfoTransform
+import chisel3.internal.sourceinfo.SourceInfoTransform
 import scala.language.experimental.macros
 import chisel3.experimental.SourceInfo
