+++ core/src/main/scala/chisel3/experimental/SourceInfo.scala
@@ -16,6 +16,7 @@ package chisel3.experimental
 
 import scala.language.experimental.macros
 import scala.reflect.macros.blackbox.Context
+import chisel3.internal.sourceinfo.SourceInfoMacro
 
 /** Abstract base class for generalized source information.
   */
@@ -45,25 +46,6 @@ case object DeprecatedSourceInfo extends NoSourceInfo
   */
 case class SourceLine(filename: String, line: Int, col: Int) extends SourceInfo {
   def makeMessage(f: String => String): String = f(s"@[$filename $line:$col]")
-}
-
-/** Provides a macro that returns the source information at the invocation point.
-  */
-@deprecated("Public APIs in chisel3.internal are deprecated", "Chisel 3.6")
-object SourceInfoMacro {
-  def generate_source_info(c: Context): c.Tree = {
-    import c.universe._
-    val p = c.enclosingPosition
-
-    val userDir = sys.props.get("user.dir") // Figure out what to do if not provided
-    val projectRoot = sys.props.get("chisel.project.root")
-    val root = projectRoot.orElse(userDir)
-
-    val path = root.map(r => p.source.file.canonicalPath.stripPrefix(r)).getOrElse(p.source.file.name)
-    val pathNoStartingSlash = if (path(0) == '/') path.tail else path
-
-    q"_root_.chisel3.experimental.SourceLine($pathNoStartingSlash, ${p.line}, ${p.column})"
-  }
 }
 
 object SourceInfo {
