+++ core/src/main/scala/chisel3/experimental/hierarchy/core/Definition.scala
@@ -7,7 +7,7 @@ import chisel3._
 
 import scala.collection.mutable.HashMap
 import chisel3.internal.{Builder, DynamicContext}
-import chisel3.internal.sourceinfomacros.{DefinitionTransform, DefinitionWrapTransform}
+import chisel3.internal.sourceinfo.{DefinitionTransform, DefinitionWrapTransform}
 import chisel3.experimental.{BaseModule, SourceInfo}
 import chisel3.experimental.hierarchy.Definition
 import firrtl.annotations.{IsModule, ModuleTarget, NoTargetAnnotation}
