+++ src/main/scala/chisel3/util/BitPat.scala
@@ -5,7 +5,7 @@ package chisel3.util
 import scala.language.experimental.macros
 import chisel3._
 import chisel3.experimental.SourceInfo
-import chisel3.internal.sourceinfomacros.SourceInfoTransform
+import chisel3.internal.sourceinfo.SourceInfoTransform
 import scala.collection.mutable
 import scala.util.hashing.MurmurHash3
