+++ src/main/scala/chisel3/util/Bitwise.scala
@@ -7,7 +7,7 @@ package chisel3.util
 
 import chisel3._
 import chisel3.experimental.SourceInfo
-import chisel3.internal.sourceinfomacros.SourceInfoTransform
+import chisel3.internal.sourceinfo.SourceInfoTransform
 import scala.language.experimental.macros
 
 /** Creates repetitions of each bit of the input in order.
