+++ src/main/scala/chisel3/util/Reg.scala
@@ -6,7 +6,7 @@ import scala.language.experimental.macros
 
 import chisel3._
 import chisel3.experimental.SourceInfo
-import chisel3.internal.sourceinfomacros.SourceInfoTransform
+import chisel3.internal.sourceinfo.SourceInfoTransform
 
 object RegEnable {
